backend/requirements.txt
# Core
pydantic-settings
ollama
pillow
pandas
pyarrow

# Image Generation
diffusers
transformers
accelerate
sentencepiece
torch
google-generativeai

# Utilities
numpy
requests
peft

# FastAPI
fastapi
uvicorn
python-multipart

# CORS & extras
fastapi[standard]
uvicorn[standard]

# Database
SQLAlchemy
psycopg2-binary

# gRPC
grpcio
grpcio-tools
protobuf
